Move `crypto` under the `util` folder
