+++ src/login/keys.js
@@ -1,4 +1,4 @@
-import { encrypt, hmacSha256 } from '../crypto/crypto.js'
+import { encrypt, hmacSha256 } from '../util/crypto/crypto.js'
 import { base16, base64, utf8 } from '../util/encoding.js'
 
 /**
