+++ src/storage/repoFolder.js
@@ -1,4 +1,4 @@
-import { decrypt, encrypt } from '../crypto/crypto.js'
+import { decrypt, encrypt } from '../util/crypto/crypto.js'
 import { utf8 } from '../util/encoding.js'
 
 /**
