+++ src/util/crypto/crypto.js
@@ -1,4 +1,4 @@
-import { base16, base64 } from '../util/encoding.js'
+import { base16, base64 } from '../encoding.js'
 import { hashjs } from './external.js'
 import aesjs from 'aes-js'
