+++ src/util/crypto/crypto.test.js
@@ -1,23 +1,8 @@
-import { makeFakeIos } from '../indexABC.js'
-import { base16, base64, utf8 } from '../util/encoding.js'
+import { makeFakeIos } from '../../indexABC.js'
+import { base16, utf8 } from '../encoding.js'
 import { decrypt, encrypt, hmacSha256, sha256 } from './crypto.js'
-import { scrypt } from './scrypt.js'
 import { assert } from 'chai'
 import { describe, it } from 'mocha'
-
-describe('scrypt', function () {
-  it('match a known userId', function () {
-    const password = utf8.parse('william test')
-    const salt = base16.parse(
-      'b5865ffb9fa7b3bfe4b2384d47ce831ee22a4a9d5c34c7ef7d21467cc758f81b'
-    )
-    const result = 'TGnly9w3Fch7tyJVO+0MWLpvlbMGgWODf/tFlNkV6js='
-
-    return scrypt(password, salt, 16384, 1, 1, 32).then(userId => {
-      return assert.equal(base64.stringify(userId), result)
-    })
-  })
-})
 
 describe('encryption', function () {
   it('decrypt existing data', function () {
