+++ src/util/crypto/external.js
@@ -1,3 +1,3 @@
-import bundle from '../../build/crypto-bundle.js'
+import bundle from '../../../build/crypto-bundle.js'
 export const elliptic = bundle.elliptic
 export const hashjs = bundle.hashjs
