+++ src/util/crypto/scrypt.js
