+++ src/util/crypto/scrypt.test.js
@@ -1,0 +1,17 @@
+import { base16, base64, utf8 } from '../encoding.js'
+import { scrypt } from './scrypt.js'
+import { assert } from 'chai'
+import { describe, it } from 'mocha'
+
+describe('scrypt', function () {
+  it('match a known userId', async function () {
+    const password = utf8.parse('william test')
+    const salt = base16.parse(
+      'b5865ffb9fa7b3bfe4b2384d47ce831ee22a4a9d5c34c7ef7d21467cc758f81b'
+    )
+    const result = 'TGnly9w3Fch7tyJVO+0MWLpvlbMGgWODf/tFlNkV6js='
+
+    const userId = await scrypt(password, salt, 16384, 1, 1, 32)
+    assert.equal(base64.stringify(userId), result)
+  })
+})
