+++ webpack.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  entry: './src/crypto/bundle.js',
+  entry: './src/util/crypto/bundle.js',
   output: {
     filename: './build/crypto-bundle.js',
     libraryTarget: 'commonjs'
